+++ Src/DeUrgenta.Services.Emailing/Constants/EmailConstants.cs
@@ -1,26 +1,27 @@
 ﻿using System.Collections.Generic;
+using DeUrgenta.Services.Emailing.Models;
 
-namespace DeUrgenta.User.Api.Services.Emailing
+namespace DeUrgenta.Services.Emailing.Constants
 {
     public static class EmailConstants
     {
         public static string GetSubject(EmailTemplate template)
         {
-            return subjects[template];
+            return Subjects[template];
         }
 
         public static string GetTemplatePath(EmailTemplate template)
         {
-            return templatePaths[template];
+            return TemplatePaths[template];
         }
 
-        private static Dictionary<EmailTemplate, string> subjects = new Dictionary<EmailTemplate, string>
+        private static readonly Dictionary<EmailTemplate, string> Subjects = new()
         {
             { EmailTemplate.AccountConfirmation, "Bine ai venit!"},
             { EmailTemplate.ResetPassword, "Resetare parolă"}
         };
 
-        private static Dictionary<EmailTemplate, string> templatePaths = new Dictionary<EmailTemplate, string>
+        private static readonly Dictionary<EmailTemplate, string> TemplatePaths = new()
         {
             { EmailTemplate.AccountConfirmation, "accountConfirmationTemplate.html"},
             { EmailTemplate.ResetPassword, "resetPasswordTemplate.html"}
