+++ Src/DeUrgenta.Services.Emailing/Models/EmailAttachment.cs
@@ -1,4 +1,4 @@
-﻿namespace DeUrgenta.User.Api.Services.Emailing
+﻿namespace DeUrgenta.Services.Emailing.Models
 {
     public class EmailAttachment
     {
