+++ Src/DeUrgenta.Services.Emailing/Models/EmailRequestModel.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 
-namespace DeUrgenta.User.Api.Services.Emailing
+namespace DeUrgenta.Services.Emailing.Models
 {
     public class EmailRequestModel
     {
