+++ Src/DeUrgenta.Services.Emailing/Models/EmailTemplate.cs
@@ -1,4 +1,4 @@
-﻿namespace DeUrgenta.User.Api.Services.Emailing
+﻿namespace DeUrgenta.Services.Emailing.Models
 {
     public enum EmailTemplate
     {
