+++ Src/DeUrgenta.Services.Emailing/Services/IEmailBuilderService.cs
@@ -1,6 +1,7 @@
 ﻿using System.Threading.Tasks;
+using DeUrgenta.Services.Emailing.Models;
 
-namespace DeUrgenta.User.Api.Services.Emailing
+namespace DeUrgenta.Services.Emailing.Services
 {
     public interface IEmailBuilderService
     {
