+++ Src/DeUrgenta.User.Api/Controller/AuthController.cs
@@ -4,13 +4,13 @@ using System.Text;
 using System.Text.Encodings.Web;
 using System.Threading.Tasks;
 using CSharpFunctionalExtensions;
+using DeUrgenta.Services.Emailing.Models;
 using DeUrgenta.User.Api.Models;
 using DeUrgenta.User.Api.Models.DTOs.Requests;
 using DeUrgenta.User.Api.Models.DTOs.Responses;
 using DeUrgenta.User.Api.Notifications;
 using DeUrgenta.User.Api.Queries;
 using DeUrgenta.User.Api.Services;
-using DeUrgenta.User.Api.Services.Emailing;
 using MediatR;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
