+++ Src/DeUrgenta.User.Api/DeUrgenta.User.Api.csproj
@@ -17,16 +17,13 @@
 		<PackageReference Include="Microsoft.AspNetCore.Identity.UI" Version="5.0.2" />
 		<PackageReference Include="IdentityServer4.AccessTokenValidation" Version="3.0.1" />
 		<PackageReference Include="Microsoft.AspNetCore.Authentication.JwtBearer" Version="5.0.4" />
-
-    <PackageReference Include="SendGrid" Version="9.13.1" />
-
-    <PackageReference Include="MailKit" Version="2.5.2" />
 	</ItemGroup>
 
 	<ItemGroup>
 		<ProjectReference Include="..\DeUrgenta.Common\DeUrgenta.Common.csproj" />
 		<ProjectReference Include="..\DeUrgenta.Domain\DeUrgenta.Domain.csproj" />
 		<ProjectReference Include="..\DeUrgenta.Infra\DeUrgenta.Infra.csproj" />
+		<ProjectReference Include="..\DeUrgenta.Services.Emailing\DeUrgenta.Services.Emailing.csproj" />
 	</ItemGroup>
 
 	<ItemGroup>
